split actual service from event listener
