+++ src/Resources/config/services.xml
@@ -3,7 +3,10 @@
            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
+        <service id="goetas.multipart_upload.request_handler" class="Goetas\MultipartUploadBundle\Service\MultipartRequestHandler">
+        </service>
         <service id="goetas.multipart_upload.request_listener" class="Goetas\MultipartUploadBundle\EventListener\MultipartRequestListener">
+            <argument type="service" key="$requestHandler" id="goetas.multipart_upload.request_handler"/>
             <tag name="kernel.event_listener" event="kernel.request" method="onKernelRequest" priority="200"/>
         </service>
     </services>
